# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QRouter API",
    "version": "1.0.0",
    "description": "One authenticated API for hardware-aware compilation, routing, pricing, and execution across quantum providers."
  },
  "servers": [
    {
      "url": "https://api.qrouter.dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "qci_live_*"
      }
    },
...
    "schemas": {
      "Constraints": {
        "type": "object",
        "properties": {
          "maxCost": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "maxQueueSeconds": {
            "type": "integer",
            "minimum": 0
          },
          "minFidelity": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "kind": {
            "enum": [
              "qpu",
              "simulator"
            ]
          },
          "providers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludeProviders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateJob": {
        "type": "object",
        "required": [
          "circuit"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "circuit": {
            "type": "string",
            "maxLength": 256000
          },
          "format": {
            "enum": [
              "openqasm2",
              "openqasm3"
            ],
            "default": "openqasm2"
          },
          "shots": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000,
            "default": 1024
          },
          "target": {
            "type": "string",
            "default": "auto"
          },
          "routing_mode": {
            "enum": [
              "balanced",
              "cost",
              "speed",
              "quality"
            ],
            "default": "balanced"
          },
          "optimization_level": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3,
            "default": 2
          },
          "failover": {
            "type": "boolean",
            "default": true,
            "description": "Try another compatible backend after a provider failure without exceeding the accepted quote."
          },
          "max_attempts": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "default": 3
          },
          "timeout_seconds": {
            "type": "integer",
            "minimum": 60,
            "maximum": 604800,
            "default": 7200
          },
          "constraints": {
            "$ref": "#/components/schemas/Constraints"
          }
        }
      },
      "ProjectSettings": {
        "type": "object",
        "required": [
          "shots",
          "target",
          "routingMode",
          "optimizationLevel"
        ],
        "properties": {
          "shots": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000,
            "default": 1024
          },
          "target": {
            "type": "string",
            "default": "auto"
          },
          "routingMode": {
            "enum": [
              "balanced",
              "cost",
              "speed",
              "quality"
            ],
            "default": "balanced"
          },
          "optimizationLevel": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3,
            "default": 2
          },
          "failover": {
            "type": "boolean",
            "default": true
          },
          "maxAttempts": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "default": 3
          },
          "timeoutSeconds": {
            "type": "integer",
            "minimum": 60,
            "maximum": 604800,
            "default": 7200
          }
        }
      },
...
      "RepositoryDeployment": {
        "type": "object",
        "required": [
          "project_id"
        ],
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "ref": {
            "type": "string",
            "description": "Branch, tag, or commit SHA. Defaults to the production branch."
          },
          "circuit_path": {
            "type": "string",
            "description": "OpenQASM entrypoint. Defaults to the project entrypoint."
          },
          "deployment_id": {
            "type": "string",
            "maxLength": 120,
            "description": "Stable idempotency value from CI or a webhook."
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettings"
          }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "created",
              "analyzing",
              "quoted",
              "awaiting_payment",
              "funds_reserved",
              "queued",
              "dispatching",
              "submitted",
              "processing",
              "completed",
              "failed",
              "cancellation_requested",
              "cancelled"
            ]
          },
          "selected_backend_id": {
            "type": "string"
          },
          "analysis": {
            "type": "object"
          },
          "route_decision": {
            "type": "object"
          },
          "result": {
            "type": [
              "object",
              "null"
            ]
          },
          "quote": {
            "type": "object",
            "nullable": true,
            "description": "Quote issued for this job (present on creation and detail responses).",
            "properties": {
              "providerCost": {
                "type": "number"
              },
              "transpilerFee": {
                "type": "number"
              },
              "platformFee": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "currency": {
                "type": "string"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              },
              "rateSnapshot": {
                "type": "object"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/repository-jobs": {
      "get": {
        "summary": "List deployments for a repository project",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository deployment list"
          }
        }
      },
      "post": {
        "summary": "Fetch a commit-pinned repository circuit, compile it, route it, and execute it",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryDeployment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deployment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Project or source not found"
          },
          "422": {
            "description": "Circuit or route is invalid"
          }
        }
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "summary": "List jobs",
        "responses": {
          "200": {
            "description": "Job list"
          }
        }
      },
      "post": {
        "summary": "Compile, quote, route, and durably queue a circuit",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJob"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Demo job created"
          },
          "202": {
            "description": "Job accepted into the durable execution queue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient credits"
          },
          "422": {
            "description": "Invalid circuit or no compatible backend"
          },
          "503": {
            "description": "Compiler unavailable"
          }
        }
      }
    },
...
  }
}